bai3/btxulihinhanh/requirements.txt
numpy
scikit-learn
opencv-python
Pillow
